Major refactor to user build system
User now uses essentially the same makefile as the kernel, instead of a
badly hacked-together minimal thing, which means it actually had real
warnings and stuff now.

It's not perfect, but is good enough for a v1.

It also looks like my gitignore was keeping a couple of the user
programs out of the repository, so they're being pulled in here as well.

Finally, I no longer have any intention of sharing libc code between
user and kernel mode, so I've moved libc/ to user/libc/.

+++ user/libc/string.c
@@ -134,7 +134,7 @@ void *memmove(void *dest_, const void *src_, size_t count) {
         // overlap, src is lower.
         // move in reverse
 
-        for (size_t i=count-1; i>=0; i--) {
+        for (long i=count-1; i>=0; i--) {
             dest[i] = src[i];
         }
         return dest;
